+++ seminar7-task50-fibo2/Program.cs
@@ -23,7 +23,7 @@ void PrintResult(string line)
 }
 
 // Метод, печатает одномерный массив, выделяет элемент с указанной позицией
-void Print2DArrColorElement(long[,] arr, int elementRow = -1, int elementCol = -1)
+void Print2DArrColorElement(double[,] arr, int elementRow = -1, int elementCol = -1)
 {
     for (int i = 0; i < arr.GetLength(0); i++)
     {
@@ -43,14 +43,14 @@ void Print2DArrColorElement(long[,] arr, int elementRow = -1, int elementCol = -
 }
 
 // Метод генерации и заполнения двумерного массива числами Фибоначчи
-long[,] Fill2DArrayFibo(int rows, int cols)
+double[,] Fill2DArrayFibo(int rows, int cols)
 {
     // Получаем массив чисел Фибоначчи нужной длины
-    long[] fiboArr = FibonacciArray(rows * cols);
+    double[] fiboArr = FibonacciArray(rows * cols);
     int fiboIndex = 0;
 
     // Создаём массив
-    long[,] array2D = new long[rows, cols];
+    double[,] array2D = new double[rows, cols];
 
     for (int i = 0; i < rows; i++)
     {
@@ -65,10 +65,10 @@ long[,] Fill2DArrayFibo(int rows, int cols)
 }
 
 // Метод генерации одномерного массива числами Фибоначчи нужной длины
-long[] FibonacciArray(int number)
+double[] FibonacciArray(int number)
 {
     // Создаём новый массив
-    long[] fibonacciArr = new long[number];
+    double[] fibonacciArr = new double[number];
     // Если длина больше 1, то во второй элемент кладём 1
     if (number > 1) fibonacciArr[1] = 1;
 
@@ -82,7 +82,7 @@ long[] FibonacciArray(int number)
 }
 
 // Метод, принимает массив и позицию, печатает результат
-void PrintColorElement(long[,] arr, int row, int col)
+void PrintColorElement(double[,] arr, int row, int col)
 {
     // Проверяем вхождение позиции элемента в массив 
     bool ifArrayHasPosition = row > 0 && row <= arr.GetLength(0) &&
@@ -100,7 +100,7 @@ void PrintColorElement(long[,] arr, int row, int col)
 
 int inputRows = ReadData("Введите количество строк: ");
 int inputCols = ReadData("Введите количество столбцов: ");
-long[,] array2D = Fill2DArrayFibo(inputRows, inputCols);
+double[,] array2D = Fill2DArrayFibo(inputRows, inputCols);
 PrintResult($"Массив чисел Фибоначчи {inputRows}x{inputCols}:");
 Print2DArrColorElement(array2D);
 Console.WriteLine();
